# src/main/java/com/users/usermanaging/dto/UserSystemDataDto.java
package com.users.usermanaging.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@AllArgsConstructor
@NoArgsConstructor
@Getter
@Setter
public class UserSystemDataDto {
    public UserSystemDTO userData;
}